+++ src/aux-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # AUX Server Changelog
 
+## V0.3.24
+### Date: TBD
+
+### Changes:
+- Bug Fixes
+  - Use context.z position has an offset from the calculated display z position in Aux Builder.
+    - Making context.z act as an offset allows context.z value of 0 to place the file on the “ground” regardless of tile height in Aux Builder and always place the file on the ground in Aux Builder.
+    - No more file clipping issues due to grid planes being at different heights between Aux Builder and Aux Player.
 ## V0.3.23
 ### Date: 3/26/2019